Refactored in two separate files. Also, unit tests are separated.
