+++ src/Humanizer/Localisation/NumberToWords/SerbianCyrlNumberToWordsConverter.cs
@@ -6,14 +6,14 @@ using System.Text;
 
 namespace Humanizer.Localisation.NumberToWords
 {
-    internal class SerbianNumberToWordsConverter : GenderlessNumberToWordsConverter
+	internal class SerbianCyrlNumberToWordsConverter : GenderlessNumberToWordsConverter
     {
-        private static readonly string[] UnitsMap = { "nula", "jedan", "dva", "tri", "četiri", "pet", "šest", "sedam", "osam", "devet", "deset", "jedanaest", "dvanaest", "trinaest", "četrnaestt", "petnaest", "šestnaest", "sedemnaest", "osemnaest", "devetnaest" };
-        private static readonly string[] TensMap = { "nula", "deset", "dvadeset", "trideset", "četrdeset", "petdeset", "šestdeset", "sedamdeset", "osamdeset", "devetdeset" };
+        private static readonly string[] UnitsMap = { "нула", "један", "два", "три", "четири", "пет", "шест", "седам", "осам", "девет", "десет", "једанест", "дванаест", "тринаест", "четрнаест", "петнаест", "шестнаест", "седамнаест", "осамнаест", "деветнаест" };
+        private static readonly string[] TensMap = { "нула", "десет", "двадесет", "тридесет", "четрдесет", "петдесет", "шестдесет", "седамдесет", "осамдесет", "деветдесет" };
 
         private readonly CultureInfo _culture;
 
-        public SerbianNumberToWordsConverter(CultureInfo culture)
+		public SerbianCyrlNumberToWordsConverter(CultureInfo culture)
         {
             _culture = culture;
         }
@@ -22,7 +22,7 @@ namespace Humanizer.Localisation.NumberToWords
         {
             if (number == 0)
             {
-                return "nula";
+                return "нула";
             }
 
             if (number < 0)
@@ -35,7 +35,7 @@ namespace Humanizer.Localisation.NumberToWords
 
             if (billions > 0)
             {
-				parts.Add(Part("milijarda", "dve milijarde", "{0} milijarde", "{0} milijarda", billions));
+				parts.Add(Part("милијарда", "две милијарде", "{0} милијарде", "{0} милијарда", billions));
                 number %= 1000000000;
                 
                 if (number > 0)
@@ -48,7 +48,7 @@ namespace Humanizer.Localisation.NumberToWords
 
             if (millions > 0)
             {
-				parts.Add(Part("milion", "dva miliona", "{0} miliona", "{0} miliona", millions));
+				parts.Add(Part("милион", "два милиона", "{0} милиона", "{0} милиона", millions));
                 number %= 1000000;
 
                 if (number > 0)
@@ -61,7 +61,7 @@ namespace Humanizer.Localisation.NumberToWords
             
             if (thousands > 0)
             {
-				parts.Add(Part("hiljadu", "dve hiljade", "{0} hiljade", "{0} hiljada", thousands));
+                parts.Add(Part("хиљаду", "две хиљаде", "{0} хиљаде", "{0} хиљада", thousands));
                 number %= 1000;
 
                 if (number > 0)
@@ -74,7 +74,8 @@ namespace Humanizer.Localisation.NumberToWords
 
             if (hundreds > 0)
             {
-				parts.Add(Part("sto", "dvesto", "{0}sto", "{0}sto", hundreds));
+               
+                parts.Add(Part("сто", "двесто", "{0}сто", "{0}сто", hundreds));
                 number %= 100;
 
                 if (number > 0)
@@ -127,5 +128,4 @@ namespace Humanizer.Localisation.NumberToWords
             }
         }
     }
-    
 }
